+++ submission/tests/test_helpers.py
@@ -263,15 +263,17 @@ def test_get_sender_email_address_zendesk_action(zendesk_action_payload):
     assert email == 'zendesk-user@example.com'
 
 
-def test_get_sender_email_address_notify_action(
-        gov_notify_email_action_payload
-):
-    email = helpers.get_sender_email_address(
-        gov_notify_email_action_payload['meta']
-    )
+def test_get_sender_email_address_notify_action(gov_notify_email_action_payload):
+    del gov_notify_email_action_payload['meta']['sender']
+    email = helpers.get_sender_email_address(gov_notify_email_action_payload['meta'])
     assert email == 'notify-user@example.com'
 
 
 def test_get_sender_email_address_pardot_action(pardot_action_payload):
     email = helpers.get_sender_email_address(pardot_action_payload['meta'])
     assert email is None
+
+
+def test_get_sender_email_address_sender(gov_notify_email_action_payload):
+    email = helpers.get_sender_email_address(gov_notify_email_action_payload['meta'])
+    assert email == 'erp+testform@jhgk.com'