+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Implemented enhancements
 - TT-2199 add testapi endpoint to delete submissions created by automated tests
 - TT-2204 add testapi endpoint to delete senders created by automated tests
+-  No ticket - Refactor makefile and speed up tests
 
 
 ## [2019.11.21](https://github.com/uktrade/directory-forms-api/releases/tag/2019.11.21_2)