+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## Pre release
 
 ### Implemented enhancements
-- TT-2199 add testapi endpoint to delete submissions created by automated tests
-- TT-2204 add testapi endpoint to delete senders created by automated tests
+- TT-2199 - add testapi endpoint to delete submissions created by automated tests
+- TT-2204 - add testapi endpoint to delete senders created by automated tests
+
+### Fixed bugs:
+- TT-2228 - Use sender email address provded by meta.sender
 
 
 ## [2019.11.21](https://github.com/uktrade/directory-forms-api/releases/tag/2019.11.21_2)